+++ docker/entrypoint.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+/home/appuser/.webdrivers/chromedriver --port=3000 --headless --whitelisted-ips="" --allowed-origins="*" --disable-dev-shm-usage --verbose